+++ mycroft/enclosure/display_manager.py
