+++ mycroft/enclosure/mark1/version.txt
